+++ packages/xc-sdk/src/platforms/substrate/types.ts
@@ -1,5 +1,7 @@
 import { Asset } from '@galacticcouncil/xc-core';
 
+import { TxEvent } from 'polkadot-api';
+
 import { Call, DryRunResult } from '../types';
 
 export interface SubstrateCall extends Call {
@@ -14,4 +16,10 @@ export interface SubstrateDryRunResult extends DryRunResult {
   error: string | undefined;
   events: any | undefined;
   xcm: any | undefined;
+}
+
+export interface SubstrateTxObserver {
+  onTransactionSend: (hash: string) => void;
+  onFinalized: (event: TxEvent) => void;
+  onError: (error: unknown) => void;
 }