+++ packages/descriptors/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @galacticcouncil/descriptors
 
+## 1.14.0
+
+### Minor Changes
+
+- c9797aa support v45.0.0 types
+
+## 1.13.1
+
+### Patch Changes
+
+- 69124d1: whitelist proxy pallet
+
 ## 1.13.0
 
 ### Minor Changes