+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @galacticcouncil/sdk
 
+## 10.8.0
+
+### Minor Changes
+
+- 6abf00c: slip fee support
+
+## 10.7.1
+
+### Patch Changes
+
+- 7ff3934: fix the H20 -> tokens dca bug
+
 ## 10.7.0
 
 ### Minor Changes